Move general functions into a helper utils file
